Pop-up delete & Img profile ok

+++ src/Components/Spot/CommComponent/TrashButton/TrashButton.tsx
@@ -14,24 +14,46 @@ export default function TrashButton({ commentId, onCommentDeleted }: TrashButton
   const [isHovering, setIsHovering] = useState(false);
 
   /**
-   * Handles the deletion of a comment.
+   * Handles the deletion of a comment after confirmation.
    * 
-   * @returnsA promise that resolves when the comment is successfully deleted.
+   * @returns A promise that resolves when the comment is successfully deleted.
    */
   const handleDelete = async (): Promise<void> => {
-    try {
-      const token = localStorage.getItem('userToken');
-      if (!token) {
-        console.error('Token manquant');
-        return;
+    const result = await Swal.fire({
+      title: 'Êtes-vous sûr?',
+      text: "Vous ne pourrez pas revenir en arrière!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Oui, supprimez-le!'
+    });
+
+    if (result.isConfirmed) {
+      try {
+        const token = localStorage.getItem('userToken');
+        if (!token) {
+          console.error('Token manquant');
+          return;
+        }
+        await axios.delete(`${API_BASE_URL}/secure/comment/${commentId}`, {
+          headers: { Authorization: `Bearer ${token}` },
+        });
+        // Call the callback function after successful deletion
+        onCommentDeleted();
+        Swal.fire(
+          'Supprimé!',
+          'Votre commentaire a été supprimé.',
+          'success'
+        );
+      } catch (error) {
+        console.error('Erreur lors de la suppression', error);
+        Swal.fire(
+          'Erreur!',
+          'Une erreur est survenue lors de la suppression.',
+          'error'
+        );
       }
-      await axios.delete(`${API_BASE_URL}/secure/comment/${commentId}`, {
-        headers: { Authorization: `Bearer ${token}` },
-      });
-      // Call the callback function after successful deletion
-      onCommentDeleted();
-    } catch (error) {
-      console.error('Erreur lors de la suppression', error);
     }
   };
